backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
Pillow==10.2.0
django-filter==23.5
google-generativeai==0.3.2
python-magic==0.4.27
whitenoise==6.6.0
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==1.1.0